# advanced/api/web/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u63a5\u53e3\u6587\u6863",
        "description": "\u63a5\u53e3\u76f8\u5173\u7684Swagger\u6587\u6863",
        "version": "1.0"
    },
    "host": "localhost/github/educate/advanced/api/web",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u7528\u6237\u5217\u8868",
                "description": "\u83b7\u53d6\u7528\u6237\u5217\u8868\u8868\uff08\u6570\u636e\u4e2d\u7684\u5b57\u6bb5\u5747\u53ef\u7528\u4f5c\u7b5b\u9009\u548c\u6392\u5e8f\uff09",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "\u5206\u9875",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "\u67e5\u8be2\u6570\u91cf",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "sex",
                        "in": "query",
                        "description": "\u6027\u522b",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "\u56fd\u5bb6",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "\u57ce\u5e02",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "return": {
                        "description": "\u7528\u6237\u5217\u8868"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "\u7528\u6237\u6ce8\u518c",
                "description": "\u521b\u5efa\u65b0\u7528\u6237",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "\u6ce8\u518c\u90ae\u7bb1",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "firstname",
                        "in": "formData",
                        "description": "\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastname",
                        "in": "formData",
                        "description": "\u59d3\u6c0f",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "sex",
                        "in": "formData",
                        "description": "\u6027\u522b(1:\u7537\uff0c0:\u5973)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "avatar",
                        "in": "formData",
                        "description": "\u5934\u50cf(URL)",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "\u7535\u8bdd",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "description": "\u56fd\u5bb6",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "description": "\u57ce\u5e02",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "adderss_1",
                        "in": "formData",
                        "description": "\u5730\u57401",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "adderss_2",
                        "in": "formData",
                        "description": "\u5730\u57402",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "return": {
                        "description": "\u7528\u6237\u4fe1\u606f"
                    }
                }
            }
        },
        "/v1/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u7528\u6237\u5217\u8868",
                "description": "\u83b7\u53d6\u7528\u6237\u5217\u8868\u8868\uff08\u6570\u636e\u4e2d\u7684\u5b57\u6bb5\u5747\u53ef\u7528\u4f5c\u7b5b\u9009\u548c\u6392\u5e8f\uff09",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u7528\u6237ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "return": {
                        "description": "\u7528\u6237\u5217\u8868"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "\u4fee\u6539\u7528\u6237\u4fe1\u606f",
                "description": "\u4fee\u6539\u7528\u6237\u4fe1\u606f",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u7528\u6237ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "firstname",
                        "in": "formData",
                        "description": "\u540d\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastname",
                        "in": "formData",
                        "description": "\u59d3\u6c0f",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "sex",
                        "in": "formData",
                        "description": "\u6027\u522b(1:\u7537\uff0c0:\u5973)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "avatar",
                        "in": "formData",
                        "description": "\u5934\u50cf(URL)",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "\u7535\u8bdd",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "description": "\u56fd\u5bb6",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "description": "\u57ce\u5e02",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "adderss_1",
                        "in": "formData",
                        "description": "\u5730\u57401",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "adderss_2",
                        "in": "formData",
                        "description": "\u5730\u57402",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "return": {
                        "description": "\u7528\u6237\u4fe1\u606f"
                    }
                }
            }
        },
        "/v1/users/favorites": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u7528\u6237\u6536\u85cf",
                "description": "\u83b7\u53d6\u7528\u6237\u6536\u85cf\u8bfe\u7a0b\u5217\u8868",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "return": {
                        "description": "\u6536\u85cf\u5217\u8868"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "\u6dfb\u52a0\u7528\u6237\u6536\u85cf",
                "description": "\u6dfb\u52a0\u7528\u6237\u6536\u85cf",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "\u7528\u6237ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "course_id",
                        "in": "formData",
                        "description": "\u8bfe\u7a0bID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "return": {
                        "description": "\u7528\u6237\u4fe1\u606f"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "\u79fb\u9664\u7528\u6237\u6536\u85cf",
                "description": "\u79fb\u9664\u7528\u6237\u6536\u85cf",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "\u7528\u6237ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "course_id",
                        "in": "formData",
                        "description": "\u8bfe\u7a0bID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "return": {
                        "description": "\u7528\u6237\u4fe1\u606f"
                    }
                }
            }
        }
    },
...
}